chore(cleanup): delete redesign/, app-header/, view-toggle, matrix-board after v9 swap
Migrate three utilities still needed by matrix-simplified/ before deletion:
- useNotificationChecker → lib/use-notification-checker.ts
- SyncStatusDisplay → components/matrix-simplified/sync-status-display.tsx
- HelpDrawer → components/matrix-simplified/help-drawer.tsx

Remove ViewToggle from settings/page.tsx (v9 ships Matrix-only; ThemeToggle kept).

Delete dead code: components/redesign/, lib/redesign/, components/app-header/,
components/app-header.tsx, components/view-toggle.tsx, components/matrix-board/,
components/matrix-board.tsx, components/matrix-column.tsx,
components/dashboard/dashboard-shell-header.tsx.

Delete obsolete tests: matrix-board, matrix-column, redesign-due, matrix-page.
Remove deleted-component describe blocks from misc-components, remaining-components,
coverage-boost-ui, and final-function-push test files.

Bump 8.7.15 → 8.7.16.
