--- components/app-header.tsx
@@ -1,1 +1,0 @@
-export { AppHeader } from "@/components/app-header/index";
