+++ components/matrix-simplified/sync-status-display.tsx
